requirements.txt
pdfplumber
pytesseract
pillow
python-docx
openpyxl
pandas
